Revert "fix header and carousal"
