+++ src/App.js
@@ -1,27 +1,24 @@
 import { React } from "react";
 import "./App.css";
-import Slider from "./components/Slider/Slider";
+import logo from "./images/logo.jpg";
+import Slider from "./components/Slider";
 import Services from "./components/Services";
 import Clients from "./components/Clients";
 import About from "./components/about";
 import Contact from "./components/contact";
 import Navbar from "./components/navbar";
-import HeaderImage from "./components/HeaderImage/HeaderImage";
-import { SiteWrapper } from "./App.styled";
 
 function App() {
     return (
-        // <div className="app">
-        <SiteWrapper>
-            <HeaderImage />
+        <div className="app">
+            <div className="header"><img className="brand" src={logo} alt="logo" /></div>
             <Slider /> 
             <Navbar />
             <Services />
             <Clients />
             <About />
             <Contact />
-        </SiteWrapper>
-        // </div>
+        </div>
     );
 }
 export default App;
