+++ src/App.styled.js
@@ -1,3 +1,3 @@
-import styled from "styled-components";
+import styled from 'styled-components'
 
-export const SiteWrapper = styled.div``;
+const Button = styled.button``
