--- src/components/HeaderImage/HeaderImage.styled.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const HeaderBar = styled.div`
-    width: 100%;
-    text-align: center;
-`;
